# src/main/java/com/openbanking/standing/orders/model/OBReadStandingOrder6.java
package com.openbanking.standing.orders.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadStandingOrder6
 */
@Getter
@Setter
public class OBReadStandingOrder6 implements Serializable {
	private static final long serialVersionUID = 1L;

}